pyproject.toml
[tool.poetry]
name = "wordpress-author-style-imitator"
version = "0.1.0"
description = "A tool to imitate a given author's writing style"
authors = ["Fırat Gelbal <firat.gelbal@automattic.com>"]
readme = "README.md"
packages = [
    {include = "wordpress_author_style_imitate", from = "src"},
    {include = "utils", from = "src"}
]

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
requests-oauthlib = "^2.0.0"
python-dotenv = "^1.0.1"
beautifulsoup4 = "^4.10.0"
haystack-ai = "^2.5.0"
transformers = "^4.44.2"
anthropic = "^0.34.2"
anthropic-haystack = "^1.0.0"
pydantic = "1.10.14"
sentence-transformers = "^3.0.1"
torch = "^2.5.0"
ollama-haystack = "^1.1.0"
colorama = "^0.4.6"

[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.2.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
